backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
gtts
moviepy
matplotlib
numpy